requirements.txt
Django==1.5.1
South==0.8.1
py-bcrypt==0.3
requests==1.2.3
stripe==1.9.2
wsgiref==0.1.2
